fix: move operator to `core/src` and add sample_data
