--- operators/detect_lewd_images/pyproject.toml
@@ -1,20 +1,0 @@
-[project]
-name = "feluda-detect-lewd-images"
-version = "0.0.0"
-requires-python = ">=3.10"
-dependencies = ["huggingface-hub", "tensorflow"]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.semantic_release]
-version_variable = ["pyproject.toml:project.version"]
-
-[tool.semantic_release.branches.main]
-match = "main"
-prerelease = false
-tag_format = "{name}-{version}"
-
-[tool.hatch.build.targets.wheel]
-packages = ["."]
